requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
cryptography==41.0.4
pymysql==1.0.3